Printing the elements of an array using a function pointer

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -7,7 +7,7 @@
   */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int x;
+	unsigned int x;
 
 	if (array == NULL)
 		return;
